refactor: Move the service to the module layer
