+++ src/modules/baileys/usecase/instance/delete/delete-instance.useCase.ts
