+++ src/modules/baileys/usecase/instance/delete/delete-instance.usecase.dto.ts
