+++ src/modules/baileys/usecase/instance/get-qr-code/get-qr-code.usecase.ts
@@ -1,6 +1,6 @@
 import {BaileysInstanceRepositoryInMemory} from "../../../repository/baileys-instance-repository-in-memory";
 import {checkInstance} from "../../../helpers/check-Instance";
-import {GetQrCodeUseCaseDTO} from "./get-qr-code.usecase.dto";
+import {GetQrCodeUseCaseInputDTO, GetQrCodeUseCaseOutPutDTO} from "./get-qr-code.usecase.dto";
 
 export class GetQrCodeUsecase {
     constructor(
@@ -8,9 +8,11 @@ export class GetQrCodeUsecase {
     ) {
     }
 
-    async execute(input: GetQrCodeUseCaseDTO){
+    async execute(input: GetQrCodeUseCaseInputDTO): Promise<GetQrCodeUseCaseOutPutDTO>{
         const result = await checkInstance(input.id, this.baileysManager)
 
-        return result.qr;
+        return {
+            qr: result.qr
+        };
     }
 }
